+++ data/trx/ship/games/tr3/scripts/jungle.lua
@@ -4,5 +4,5 @@ end)
 
 trx.events.before_item_setup(function(level)
   -- Setup shoals
-  trx.items[134].properties.range = { x = 10, y = 3, z = 22 }
+  trx.items[133].properties.range = { x = 10, y = 3, z = 22 }
 end)