+++ data/trx/ship/games/tr3/scripts/tonyboss.lua
@@ -1,5 +1,5 @@
 trx.events.on_pickup(function(pickup_item)
-  local item = trx.items[pickup_item + 1]
+  local item = trx.items[pickup_item]
   if item.object_id == trx.catalog.objects.quest_item_1 then
     trx.rooms.flip_effect(trx.catalog.flip_effects.finish_level)
   end