+++ data/trx/ship/games/tr4/scripts/cortyard.lua
@@ -1,3 +1,3 @@
 trx.events.before_item_setup(function(level)
-  trx.items[2].properties.lift = true
+  trx.items[1].properties.lift = true
 end)