+++ data/trx/ship/games/tr4/scripts/library.lua
@@ -1,4 +1,4 @@
 trx.events.before_item_setup(function(level)
-  trx.items[13].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
-  trx.items[136].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
+  trx.items[12].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
+  trx.items[135].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
 end)