+++ data/trx/ship/games/tr4/scripts/lowstrt.lua
@@ -1,5 +1,5 @@
 trx.events.before_item_setup(function(level)
-  trx.items[46].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
+  trx.items[45].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
 
   trx.objects.animating_13.properties.collidable = false
   trx.objects.animating_14.properties.collidable = false