+++ data/trx/ship/games/tr4/scripts/settomb1.lua
@@ -1,5 +1,5 @@
 trx.events.before_item_setup(function(level)
+  trx.items[3].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
   trx.items[4].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
-  trx.items[5].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
-  trx.items[88].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
+  trx.items[87].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
 end)